Refresh dependencies to latest stable (Kotlin 2.3, Compose BOM 2026.06, Health Connect 1.1.0) (#54)
* Refresh dependencies to latest stable (Kotlin 2.3, Compose BOM 2026.06, Health Connect 1.1.0)

Bump gradle/libs.versions.toml to latest stable (no RC/alpha/beta, staying on AGP 8.x):

- agp 8.11.1 -> 8.13.2
- kotlin 2.2.0 -> 2.3.21 (also drives compose-compiler + kapt)
- ksp 2.2.0-1.0.29 -> 2.3.10 (tracks Kotlin 2.3.x)
- koin 4.1.0 -> 4.2.2
- coreKtx 1.16.0 -> 1.18.0 (held below 1.19.0; see below)
- lifecycle x3 2.9.2 -> 2.10.0 (held below 2.11.0; see below)
- activityCompose 1.10.1 -> 1.13.0
- navigationCompose 2.9.3 -> 2.9.8
- composeBom 2025.07.00 -> 2026.06.01
- healthConnect 1.1.0-rc03 -> 1.1.0 (RC -> stable)
- datetime 0.7.1 -> 0.8.0
- kotlinx-serialization 1.9.0 -> 1.11.0
- datastore(+preferences) 1.1.7 -> 1.2.1
- room 2.7.2 -> 2.8.4
- robolectric 4.16 -> 4.16.1
- roborazzi 1.56.0 -> 1.67.0
- mockk 1.13.12 -> 1.14.11

Compatibility code changes required by the bumps:
- Compose BOM 2026.06 no longer pulls material-icons-core transitively via
  material3; added an explicit material-icons-core dependency (BOM-managed) so
  the Icons.* references keep resolving.
- Health Connect 1.1.0 types ReadRecordsResponse.records as List<Record>, so the
  redundant `as List<Record>` casts (and their UNCHECKED_CAST suppressions) in
  HealthConnectGateway are removed.
- Health Connect 1.1.0 exposes FloorsClimbedRecord.floors as Double; dropped the
  now-redundant .toDouble() conversions.
- Compose Material3 deprecations: TabRow -> PrimaryTabRow,
  centerAlignedTopAppBarColors -> topAppBarColors.

Held back (kept below the issue's targets): coreKtx 1.19.0 and lifecycle 2.11.0
hard-require AGP 9.1.0+ / compileSdk 37. The AGP 8->9 migration is issue #49, so
these are pinned to the highest AGP-8-compatible stable (1.18.0 / 2.10.0).

Fixes #48


* ci: update screenshot baselines

---------

+++ app/src/main/java/com/monkopedia/healthdisconnect/HealthRecordMeasurementExtractor.kt
@@ -534,7 +534,7 @@ class DefaultHealthRecordMeasurementExtractor : HealthRecordMeasurementExtractor
             scalarMappingLookup(
                 fieldName = "Floors",
                 recordClass = FloorsClimbedRecord::class,
-                valueField = { it.floors.toDouble() },
+                valueField = { it.floors },
                 unitLabel = "floors"
             ),
             scalarMappingLookup(
